fix(src): remediate SonarCloud findings and replace banned strlen
Clears the SonarCloud code smells that a fresh analysis surfaced, and the
docs/SRCBANNED.md #1 violation underneath four of them:
- webdav: replace 4 unbounded strlen() with bounded strnlen(p, cap) (ban #1 /
  Sonar S5813) - a lock path/token is measured against its buffer, never scanned
  past it.
- sb_modbus: take the function address explicitly for the driver vtable pointers
  (S936).
- opcua/cotp: reword two wire-layout comments so they read as prose, not
  commented-out code (S125); drop a redundant uint32 cast in the ERR builder
  (S1905).
- ipsec_db: make the size_t->uint8_t address-length narrowing explicit (S5276).
- dnp3: split two multi-identifier declarations (S1659).
No behavior change; native dnp3/opcua/cotp/webdav/ikev2/southbound suites pass.

+++ src/services/esp/ipsec_db.cpp
@@ -93,7 +93,7 @@ bool dws_ipsec_selector_from_ts(IpsecSelector *out, const IkeTrafficSelector *ts
         return false;
 
     memset(out, 0, sizeof(*out));
-    uint8_t len = ts_src->addr_len;
+    uint8_t len = (uint8_t)ts_src->addr_len; // addr_len is an IP address length (<= 16), fits a uint8_t
     out->addr_len = len;
     out->ip_protocol = ts_src->ip_protocol ? ts_src->ip_protocol : ts_dst->ip_protocol;
     memcpy(out->src_lo, ts_src->start_addr, len);
